chore(agnostification): move Config to common dir and add path config
