+++ src/node/BrowserFetcher.ts
@@ -24,7 +24,6 @@ import * as http from 'http';
 
 import { Product } from '../common/Product.js';
 import extractZip from 'extract-zip';
-import Config from '../Config.js';
 import { promisify } from 'util';
 import removeRecursive from 'rimraf';
 import * as URL from 'url';
@@ -34,8 +33,8 @@ import createHttpsProxyAgent, {
 } from 'https-proxy-agent';
 import { getProxyForUrl } from 'proxy-from-env';
 import { assert } from '../common/assert.js';
+import debug from 'debug';
 
-const { debug } = Config;
 const debugFetcher = debug(`puppeteer:fetcher`);
 
 const downloadURLs = {
